requirements.txt
langchain_core==0.3.51
langchain_google_genai==2.1.2
langgraph==0.3.30
pydantic==2.11.3
python-dotenv==1.1.0
rich==14.0.0